Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,17 +1,17 @@
 #include <stdio.h>
 int main () {
-    double a, b;
+    int a, b;
     char operator;
-    scanf("%f %f %c", &a, &b, &operator);
+    scanf("%d %d %c", &a, &b, &operator);
     if (operator == '+') {
-        printf("%.2lf + %.2lf = %.2lf", a, b, a + b);
+        printf("%d + %d = %d", a, b, a + b);
     } else if (operator == '-') {
-        printf("%.2lf - %.2lf = %.2lf", a, b, a - b);
+        printf("%d - %d = %d", a, b, a - b);
     } else if (operator == '*') {
-        printf("%.2lf * %.2lf = %.2lf", a, b, a * b);
+        printf("%d * %.2d = %d", a, b, a * b);
     } else if (operator == '/') {
         if (b != 0) {
-            printf("%.2lf / %.2lf = %.2lf", a, b, a / b);
+            printf("%d / %d = %d", a, b, a / b);
         } else {
             printf("Division by zero not possible");
         }
